overrides merchant getstruckby
takes out casting as well

+++ Human.cc
@@ -59,7 +59,7 @@ char Human::getChar() {
 }
 
 void Human::getStruckBy(Character &other) {
- int net = (int) ceil((100.0 / (100.0 + static_cast<float>(Def))) * static_cast<float>(other.getAtk()));
+ int net = ceil((100.0 / (100.0 + static_cast<float>(Def))) * static_cast<float>(other.getAtk()));
  int amount;
 
  if (net <= 0) {
